feat(config): Refactor configuration management - Optimize JWT Auth module to load configuration via `global_args`. - Decouple configuration-related code from `utils_api.py`, and add a new `config.py` file for unified configuration management. - Adjust configuration import in `lightrag_server.py`, `auth.py`, and `document_routes.py` to be introduced through `global_args`.
